Code/requirements.txt
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
databases==0.4.1
fastapi==0.63.0
geos==0.2.2
h11==0.12.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
motor==2.3.0
paramiko==2.7.2
Pillow==8.1.0
pycparser==2.20
pydantic==1.7.3
pylint==2.6.0
pymongo==3.11.2
PyNaCl==1.4.0
python-dotenv==0.15.0
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.22
sshtunnel==0.4.0
starlette==0.13.6
toml==0.10.2
uvicorn==0.13.3
Werkzeug==1.0.1
wrapt==1.12.1
